+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 * New `exit_if_no_outdated` setting, to exit immediately if not outdated packages are found
 * New `immediate_first_check` setting, to perform first outdated-check as soon as ChocoButler starts
+* New menu entries under 'Advanced' to edit settings file and show the README file on the web
 ### Changes
 * Improved checking for a _Chocolatey GUI_ installation